change create to dropdown and enable creation of scored and adaptive pages
