+++ lib/oli/authoring/editing/util.ex
@@ -1,5 +1,6 @@
 defmodule Oli.Authoring.Editing.Utils do
   alias Oli.Accounts
+  alias Oli.Resources.Numbering
 
   def authorize_user(author, project) do
     case Accounts.can_access?(author, project) do
@@ -65,4 +66,13 @@ defmodule Oli.Authoring.Editing.Utils do
     end
   end
 
+  def new_container_name(numberings, parent_container) do
+    numbering = Map.get(numberings, parent_container.id)
+
+    if numbering do
+      Numbering.container_type(numbering.level + 1)
+    else
+      "Unit"
+    end
+  end
 end
